Added `Enumerables` static extension class
The class currently contains a `Rotate` method that will rotate a collection by a certain number of places.

The behavior is similar to `Enumerable.Skip` when too many places are rotated; if the number of places exceeds the length of the collection, the original collection is returned.  This is to avoid counting the collection.